move tests, fix size
tests moved to other package
int size -> public int size
